% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archeosciences2018-package.R
\docType{package}
\name{archeosciences2018-package}
\alias{archeosciences2018-package}
\alias{_PACKAGE}
\alias{archeosciences2018}
\title{archeosciences2018: Replication of Frerebeau, N. and Pernot, M. (2018)}
\description{
Research compendium of Frerebeau, N. and Pernot, M. 
    (2018) <doi:10.4000/archeosciences.6007> [in french]. 
    The compendium contains all data, code, and 
    text associated with this paper.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab archeosciences2018 \cr
 \strong{Type:} \tab Compendium \cr
 \strong{Version:} \tab 1.0 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.3497250}{10.5281/zenodo.3497250}
}
}
\note{
This study received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the program \emph{Investissements d'avenir}
(ref. ANR-10-LABX-52).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/nfrerebeau/archeosciences2018}
  \item Report bugs at \url{http://github.com/nfrerebeau/archeosciences2018/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France}
 Michel Pernot \tab \emph{CNRS, France}
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
